Add api to get or kill foreground app
We can kang what sysui assistant code does to check if launcher is running,
using a TaskStackChangeListener to always know the running
taskid and its component name. This is a solid and reliable way
to easily kill the task and even remove it from recents without the
risk to kill launcher or a different app than the foreground one.

Change-Id: I03279ea6139ce5a371e5266e4b35e30f11a3372f

Refactor killapp implementation into a Dependency controller class

There is a lot of good reusable code here that will be used for
future features. This makes the code universally accessable.

Change-Id: Ic0849ae363a83c9e504771eb3f025b19decfebcd

Taskhelper api: fix NPE

reported while installing an additional launcher from playstore

08-05 14:20:43.787 10550 10593 E AndroidRuntime: FATAL EXCEPTION: android.bg
08-05 14:20:43.787 10550 10593 E AndroidRuntime: Process: com.android.systemui, PID: 10550
08-05 14:20:43.787 10550 10593 E AndroidRuntime: java.lang.NullPointerException: Attempt to invoke virtual method 'java.lang.String android.content.ComponentName.getPackageName()' on a null object reference
08-05 14:20:43.787 10550 10593 E AndroidRuntime: 	at com.android.systemui.statusbar.policy.TaskHelper.isLauncherShowing(TaskHelper.java:277)
08-05 14:20:43.787 10550 10593 E AndroidRuntime: 	at com.android.systemui.statusbar.policy.TaskHelper$3.run(TaskHelper.java:145)

Hide statusbar clock when home is showing [1/2]

Many third party launchers used to offer this, but it is not
seen very ofter anymore. So provide a universal solution. Useful
when user has a clock widget on the home screen.

Change-Id: I144389c91c5c27195543514daed7c5303255101f

TaskHelper: Update callbacks on overlay change

Much of the statusbar is recreated on a overlay change so callbacks may need a refresh. Fixes clock disappearing after theme change.

Change-Id: Iae34bef2ebabb708e1d7b2d192f5e940ace80864

TaskHelper: allow to get foreground app label
Change-Id: Ic3d9d3a7e379c45573d6378fce1e6399acbb75bd
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: Jayant-Deshmukh <jayantdeshmuk008@gmail.com>
Signed-off-by: mukesh22584 <mks22584@gmail.com>