[Account][Bugfix] Fix ServiceConnection leak in AccountManagerService.
If an Activity is given, ServiceConnection implemented by session won't get
unbound until app process died, which causes leak. On the other hand,
MESSAGE_TIMED_OUT aren't sent anywhere, which seems a code missing. So
we send this message to ensure it can gets unbound.

Bug: 323308345
Test: Write an app that calls AccountManager.getAuthToken by passing
activity as parameter. You can refer to this project at github:
https://github.com/SamuGG/android-account-manager-example
Then check ServiceConnection instances in LoadedApk.mServices to see if
it is released.

Change-Id: Ib00e111f683b5bb1ca36fa910b0216e960648b49
Signed-off-by: mukesh22584 <mks22584@gmail.com>