Improve robustness when migrating from appops.xml to appops_accesses.xml
If errors are found in appopps.xml do not throw exception but log error
and return.

Rescue party is not started at the time when appops migration happens,
this means that update will be stuck in boot process and will not
attempt to reboot to recover.

User will this experience the device as bricked.

Also improve parsing of op-tag, "m" may not be written to appops if
it is the default value. If it is not present in xml, then use the
default mode for operation.

Bug: 323262109
Test: Manual
Change-Id: Ic24259662e1c7f0cfe23c72bd1c390f9aff80bec
Signed-off-by: mukesh22584 <mks22584@gmail.com>